move tf tree printing into tf_util so other applications can use it
