+++ tf_roslibrust/src/bin/tree.rs
@@ -4,52 +4,8 @@
 /// show a text version of the ros tf tree as received on /tf and /tf_static
 use clap::{arg, command};
 use roslibrust::ros1::NodeHandle;
-use std::collections::{HashMap, HashSet};
-use tf_roslibrust::{TfBuffer, TfListener};
-
-/// print the current tf tree
-/// adapted from tf_demo tf_tree.py
-
-fn print_tree(
-    tf_buffer: &TfBuffer,
-    parent_to_children: &HashMap<String, HashSet<String>>,
-    parent: &str,
-    level: u8,
-) -> Result<(), anyhow::Error> {
-    if level > 1 {
-        for _ in 0..level {
-            print!("   ");
-        }
-    }
-    if level > 0 {
-        print!("-- ");
-    }
-
-    print!("{parent}");
-    let rate = tf_buffer.get_rate(parent);
-    if let Some(rate) = rate {
-        if rate > 0.0 {
-            print!("  {rate:.2}Hz");
-        }
-    }
-    println!();
-
-    match parent_to_children.get(parent) {
-        Some(children) => {
-            let mut children = children.iter().collect::<Vec<_>>();
-            children.sort();
-            for child in children {
-                let _ = print_tree(tf_buffer, parent_to_children, child, level + 1);
-            }
-        }
-        None => {
-            // no children
-            return Ok(());
-        }
-    }
-
-    Ok(())
-}
+use tf_roslibrust::tf_util;
+use tf_roslibrust::TfListener;
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
@@ -107,21 +63,7 @@ async fn main() -> Result<(), anyhow::Error> {
         listener.force_finish();
 
         let buffer = listener.buffer.read().unwrap();
-        let parent_to_children = buffer.get_parent_to_children();
-        if false {
-            let mut keys_sorted = parent_to_children.keys().collect::<Vec<_>>();
-            keys_sorted.sort();
-            for parent in keys_sorted {
-                println!("{parent}: {:?}", parent_to_children.get(parent).unwrap());
-            }
-        }
-
-        let roots = buffer.get_roots()?;
-        for root in roots {
-            println!("[");
-            let _ = print_tree(&buffer, &parent_to_children, &root, 0);
-            println!("]");
-        }
+        let _ = tf_util::print_tree(&buffer);
     }
 
     // TODO(lucasw) wait for async unregistering
